% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_theta_mvn.R
\name{create_theta_mvn}
\alias{create_theta_mvn}
\title{A Function to Create Variables from a Multivariate Normal Distribution}
\usage{
create_theta_mvn(size, mean_vec, sd_vec, ifcor_vec)
}
\arguments{
\item{size}{A sample size, which is a positive integer value.}

\item{mean_vec}{A vector of means of the variables.}

\item{sd_vec}{A vector of standard deviation.}

\item{ifcor_vec}{This vector contains a pair-wise interfactor correlation values,
which are arranged for the pair Factors i and j, where i = 1, 2,...,m-1,
j = 2, 3,...,m, and m is the total number of dimensions. For example,
with three factors, the pair facotrs i and j is arranged as follows: Factor 1 and Factor 2,
Factor 1 and Factor 3, and Factor 2 and Factor 3.}
}
\value{
It will return two data sets: a raw data set and a scaled data set.
A scale data set has each dimension standardized. That is each dimension is
mean centered by the dimension's mean and scaled by the dimension's
standard deviation (standardized).
}
\description{
Given a vector means, standard deviation, and interfactor correlation,
create_theta_mvn() creates theta values from a multivariate normal distribution
with the number of dimensions equal to the length of the vector of
standard deviation.
}
\examples{
# example: Three-factor model
# n. of dimensions = 3
mean_vec <- c(-0.1, 0, 0.1) # mean vector for each dimension
sd_vec <- c(0.95, 0.98, 1.1) # sd vector for each dimension
f_cor <- c(0.3, 0.5, 0.6) # interfactor correlation vector
# convert the cor. vector to cor.matrix

# create theta values
# 2-multivariate normal theta with three dimensions

output <- create_theta_mvn(size = 10, 
                          mean_vec = c(-0.1, 0, 0.1), 
                          sd_vec = c(0.95, 0.98, 1.1),
                          ifcor_vec = c(0.3, 0.5, 0.6))

# Get raw theta values (latent variable scores or ability scores)
output$unscaled_ds
# Get standardized theta values (mean = 0, standard deviation =1)
output$scaled_ds

}
\seealso{
\code{\link[=to_cormatrix]{to_cormatrix()}},\code{\link[=to_covmatrix]{to_covmatrix()}}, \code{\link[=TSK]{TSK()}}
}
