% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_residuals.R
\name{diag_residuals}
\alias{diag_residuals}
\title{Compute Item Residuals from a Factor Loading Vector or Matrix}
\usage{
diag_residuals(x)
}
\arguments{
\item{x}{This is a vector of factor loadings of a p x m matrix of factor
loadings, where p is the total number of items and m is the total number of
dimensions.}
}
\value{
diag_residuals will return a p x p diagonal matrix of residuals,
in which the off diagonal cells are zero, and p is the total number of items.
}
\description{
The function diag_residuals will compute item residuals from the standardized
factor loadings.
}
\examples{
library(AUTTT)
# Example 1 
lambs <- runif(12, .7, .95)
x <- lambs
diag_residuals(x = lambs)

# Example 2
lambs <- runif(12, .7, .95)
mymodel <- list(c(1,2,3,4), c(5,6,7), c(8, 9,10, 11, 12))
to_loadingmatrix(loading_vec = lambs, model = mymodel)
x <- lambs
diag_residuals(x = lambs)
}
