% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_ML_probit.R
\name{execute_ML_probit}
\alias{execute_ML_probit}
\title{main_dir2 <- "C:/Users/shh6304/Desktop/t8"
execute_ML_probit(main_dir2 = "C:/Users/shh6304/Desktop/t8",
nReps = 500)}
\usage{
execute_ML_probit(main_dir2, nReps)
}
\description{
main_dir2 <- "C:/Users/shh6304/Desktop/t8"
execute_ML_probit(main_dir2 = "C:/Users/shh6304/Desktop/t8",
nReps = 500)
}
