% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_mplus_scripts.R
\name{execute_mplus_scripts}
\alias{execute_mplus_scripts}
\title{Execute Mplus Input Files}
\usage{
execute_mplus_scripts(
  main_dir,
  cell_folders,
  est_folders,
  start_condition,
  end_condition
)
}
\arguments{
\item{main_dir}{This is the main directory in which folders representing
study conditions and data sets for the conditions are stored.}

\item{cell_folders}{This object contains the names of the the folders representing
the experimental conditions/cells.}

\item{est_folders}{It is the folder name(s), where Mplus scripts are created and
stored using the function mplus_montecarlo_analysis_grm2().}

\item{start_condition}{It is a numeric starting value representing an index
of the cell_folders. It can range from 1 to length(cell_folders).}

\item{end_condition}{It is a numeric ending value representing an index
of the cell_folders. It can range from 1 to length(cell_folders).}
}
\description{
The function executes Mplus scripts (.inp) in the background. .
}
\examples{
library(AUTTT)

main_dir2 <- "C:/Users/shh6304/Desktop/test6"
# load folders
load(paste0(main_dir2, "/folders.Rdata"))
cell_folders <- folders
est_folder <- c("ULSMV_delta")
execute_mplus_scripts(main_dir = main_dir2, 
                     cell_folders = cell_folders, 
                     est_folders = ULSMV_delta, 
                     start_condition = 1, 
                     end_condition = length(cell_folders))
                     
}
