% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_mplus_output.R
\name{gather_mplus_output}
\alias{gather_mplus_output}
\title{Gather Mplus Output in One Table}
\usage{
gather_mplus_output(
  main_dir,
  cell_folders,
  est_folder,
  cell_prefix,
  nreps,
  methods
)
}
\arguments{
\item{cell_prefix}{It is prefix given to an R object. The function will add
a number after it to make an object name for a study condition (e.g., U1).}

\item{nreps}{It is the total of replications for a conditions/cell.}

\item{nfolders}{This object contains the names of the the folders representing
the experimental conditions/cells.}
}
\value{
It will return a table with all standardized parameter estimates
obtained by a particular estimator for all replications and conditions.
}
\description{
The function gather standardized parameter estimates produced by Mplus.
It reads the file with the extention ".out".
}
\examples{
library(AUTTT)

main_dir2 <- "C:/Users/shh6304/Desktop/test"
load(paste0(main_dir2, "/folders.Rdata"))
cell_folders <- folders
est_folder <- c("ULSMV_delta")
cell_prefix <- "U"

ulsmv_est <- gather_mplus_output(nfolders = length(cell_folders),
                                 nreps = 5,
                                 cell_prefix = "U")
}
