% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cormatrix.R
\name{to_cormatrix}
\alias{to_cormatrix}
\title{Convert a Correlation Vector to a Correlation Matrix}
\usage{
to_cormatrix(cor_vec, n_dim)
}
\arguments{
\item{cor_vec}{This is a correlation matrix that multiple pair-wise correlations in the arrangement of pairs i and j,
where i = 1,...,n_dim - 1 and j = 2,...n_dim and n_dim is the number of dimensions.}

\item{n_dim}{The number of dimensions or columns.}
}
\value{
to_cormatrix will return a n_dim x n_dim correlation matrix with 1 in the diagonal cells.
}
\description{
The function to_cormatrix converts a correlation vector to a correlation matrix
}
\examples{
to_cormatrix(cor_vec = c(0.3, 0.4, 0.5, 0.6, 0.42, 0.52), 
n_dim = 4)

}
