% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_covmatrix.R
\name{to_covmatrix}
\alias{to_covmatrix}
\title{Convert a Correlation Matrix to Covariance Matrix}
\usage{
to_covmatrix(cor_matrix, sd_vec)
}
\arguments{
\item{cor_matrix}{A a symmetric correlation matrix.}

\item{sd_vec}{A vector of the variables' standard deviations.}
}
\description{
Given a correlation matrix and correct number of standard deviations, tocovmatrix() converts the correlation matrix
to a covariance matrix. If you only have a vector correlation, you can convert it to a correlation matrix first, using
the function to_cormatrix().
}
\examples{
mycormat <- to_cormatrix(cor_vec = c(0.3, 0.5, 0.6), n_dim=3)
to_covmatrix(mycormat, sd_vec = c(0.95, 0.98, 1.1))
}
