% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_loadingmatrix.R
\name{to_loadingmatrix}
\alias{to_loadingmatrix}
\title{Convert a Correlation Vector to a Correlation Matrix}
\usage{
to_loadingmatrix(loading_vec, model)
}
\arguments{
\item{loading_vec}{This is a vector of factor loadings for each item listed
in order of item1, item2,..., itemp.}

\item{model}{is a list object specifying items in each factor in order.
See the example.}
}
\value{
to_loadingmatrix will return a p x m matrix of factor loadings
according to the confirmatory factor analysis (CFA)  model provided,
where p is the total number of items, and m
is the total number of dimensions.
}
\description{
The function to_loadingmatrix converts a vector of factor loadings
to a p x m factor loading matrix in a confirmatory factor analysis
model. This creates a loading matrix with no cross-factor loadings.
}
