% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSK.R
\name{TSK}
\alias{TSK}
\title{Thresholds, Skewness, and Kurtosis}
\usage{
TSK(n, res_prop)
}
\arguments{
\item{n}{A sample size as an interger.}

\item{res_prob}{A vector of categorical response probabilities.}
}
\value{
TSK will return threshold values and corresponding response probabilities and cumulative probabilities.
It will also return skewness and kurtosis values. Kurt1 uses the regular formula of kurtosis
but it subtract the value with -3. Kurt2 is the original value.
Kurt3 uses formula found in Sheskin (2000).
}
\description{
This package contains the TSK function {TSK()},
which allows you to calculate thresholds, skewness, and kurtosis
give categorical response probabilities.
}
\examples{
# Probability of each response category
res1<-4/120
res2<-16/120
res3<-50/120
res4<-50/120
res.prop<-c(res1, res2, res3, res4)
TSK(n = 1200, res_prop = res.prop)

}
\references{
\insertRef{Hahs_Vaughn2012}{AUTTT}

\insertRef{Finney2013}{AUTTT}

\insertRef{Kline2015}{AUTTT}

\insertRef{McDonald1999}{AUTTT}

\insertRef{UCLA}{AUTTT}
}
